Refactors models.py, adds flake8, removes BADGE_LOCK
