+++ setup.cfg
@@ -1,3 +1,5 @@
+[flake8]
+
 [pep8]
 max-line-length=999
 ignore=E121,E123,E126,E226,E24,E704,E221,E127,E128,W503,E731,E131,E711,E712
