+++ tox.ini
@@ -19,7 +19,7 @@
 #   $ tox -- -s -v -k expression
 #
 [tox]
-envlist=pep8,py34
+envlist=flake8,pep8,py34
 skipsdist=True
 
 [testenv]
@@ -40,6 +40,11 @@ commands=
 
     # clean up after ourselves
     rm -f {envdir}/src/sideboard/plugins/uber
+
+[testenv:flake8]
+deps=flake8
+commands=
+    flake8 uber/models/
 
 [testenv:pep8]
 deps=pep8
