+++ uber/common.py
@@ -79,6 +79,8 @@ from uber.utils import *
 from uber.reports import *
 from uber.decorators import *
 from uber.models import *
+from uber.models.types import *
+from uber.models.promo_code import *
 from uber.automated_emails_server import *
 from uber.automated_emails import *
 from uber.badge_funcs import *
