+++ uber/config.py
@@ -590,9 +590,6 @@ c.TRANSFERABLE_BADGE_TYPES = [getattr(c, badge_type.upper()) for badge_type in c
 
 c.DEPT_HEAD_CHECKLIST = _config['dept_head_checklist']
 
-c.BADGE_LOCK = RLock()
-c.ASSIGN_ATTENDEE_TO_GROUP_LOCK = RLock()
-
 c.CON_LENGTH = int((c.ESCHATON - c.EPOCH).total_seconds() // 3600)
 c.START_TIME_OPTS = [(dt, dt.strftime('%I %p %a')) for dt in (c.EPOCH + timedelta(hours=i) for i in range(c.CON_LENGTH))]
 c.DURATION_OPTS = [(i, '%i hour%s' % (i, ('s' if i > 1 else ''))) for i in range(1, 9)]
