+++ uber/models/__init__.py
@@ -1,0 +1,1306 @@
+import re
+import uuid
+from collections import defaultdict
+from datetime import date, datetime, timedelta
+from itertools import chain
+from uuid import uuid4
+
+import bcrypt
+import cherrypy
+import six
+import sqlalchemy
+from dateutil import parser as dateparser
+from sideboard.lib import listify, log, on_startup, stopped
+from sideboard.lib.sa import check_constraint_naming_convention, \
+    declarative_base, JSON, SessionManager, UTCDateTime, UUID
+from sqlalchemy import and_, func, or_, not_
+from sqlalchemy.event import listen
+from sqlalchemy.exc import IntegrityError
+from sqlalchemy.orm import Query, joinedload, subqueryload
+from sqlalchemy.orm.attributes import get_history, instance_state
+from sqlalchemy.schema import MetaData
+from sqlalchemy.types import Boolean, Integer, Float, Date, Numeric
+from sqlalchemy.util import immutabledict
+
+from uber.config import c
+from uber.decorators import cached_classproperty, cost_property, \
+    suffix_property
+from uber.models.types import Choice, DefaultColumn as Column, MultiChoice
+from uber.utils import check_csrf, get_real_badge_type, DeptChecklistConf
+
+
+# Consistent naming conventions are necessary for alembic to be able to
+# reliably upgrade and downgrade versions. For more details, see:
+# http://alembic.zzzcomputing.com/en/latest/naming.html
+naming_convention = {
+    'ix': 'ix_%(column_0_label)s',
+    'uq': 'uq_%(table_name)s_%(column_0_name)s',
+    'fk': 'fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s',
+    'pk': 'pk_%(table_name)s'}
+
+if not c.SQLALCHEMY_URL.startswith('sqlite'):
+    naming_convention['unnamed_ck'] = check_constraint_naming_convention
+    naming_convention['ck'] = 'ck_%(table_name)s_%(unnamed_ck)s',
+
+metadata = MetaData(naming_convention=immutabledict(naming_convention))
+
+
+@declarative_base(metadata=metadata)
+class MagModel:
+    id = Column(UUID, primary_key=True, default=lambda: str(uuid4()))
+
+    required = ()
+
+    @cached_classproperty
+    def _class_attr_names(cls):
+        return [
+            s for s in dir(cls)
+            if s not in ('_class_attrs', '_class_attr_names') and
+            not s.startswith('_cached_')]
+
+    @cached_classproperty
+    def _class_attrs(cls):
+        return {s: getattr(cls, s) for s in cls._class_attr_names}
+
+    def _invoke_adjustment_callbacks(self, label):
+        callbacks = []
+        for name, attr in self._class_attrs.items():
+            if hasattr(attr, '__call__') and hasattr(attr, label):
+                callbacks.append(getattr(self, name))
+        callbacks.sort(key=lambda f: getattr(f, label))
+        for function in callbacks:
+            function()
+
+    def presave_adjustments(self):
+        self._invoke_adjustment_callbacks('presave_adjustment')
+
+    def predelete_adjustments(self):
+        self._invoke_adjustment_callbacks('predelete_adjustment')
+
+    @property
+    def addons(self):
+        """
+        This exists only to be overridden by other events; it should return a
+        list of strings are the extra things which an attendee or group has
+        purchased.  For example, in the MAGStock codebase, we've got code which
+        looks something like this::
+
+            @Session.model_mixin
+            class Attendee:
+                purchased_food = Column(Boolean, default=False)
+
+                @property
+                def addons(self):
+                    return ['Food'] if self.purchased_food else []
+
+        Our various templates use this information to display a summary to the
+        user of what they have purchased, e.g. in the prereg confirmation page
+        and in their confirmation emails.
+        """
+        return []
+
+    @cached_classproperty
+    def cost_property_names(cls):
+        """Returns the names of all cost properties on this model."""
+        return [
+            s for s in cls._class_attr_names
+            if s != 'cost_property_names' and
+            isinstance(getattr(cls, s), cost_property)]
+
+    @cached_classproperty
+    def multichoice_columns(cls):
+        return [
+            c for c in cls.__table__.columns
+            if isinstance(c.type, MultiChoice)]
+
+    @property
+    def default_cost(self):
+        """
+        Returns the sum of all @cost_property values for this model instance.
+
+        Because things like discounts exist, we ensure default_cost will never
+        return a negative value.
+        """
+        return max(0, sum(getattr(self, s) for s in self.cost_property_names))
+
+    @property
+    def stripe_transactions(self):
+        """
+        Returns all logged Stripe transactions with this model's ID.
+        """
+        from uber.models.commerce import StripeTransaction
+        return self.session.query(StripeTransaction).filter_by(
+            fk_id=self.id).all()
+
+    @cached_classproperty
+    def unrestricted(cls):
+        """
+        Returns a set of column names which are allowed to be set by non-admin
+        attendees filling out one of the registration forms.
+        """
+        return {
+            col.name for col in cls.__table__.columns
+            if not getattr(col, 'admin_only', True)}
+
+    @cached_classproperty
+    def all_bools(cls):
+        """Returns the set of Boolean column names for this table."""
+        return {
+            col.name for col in cls.__table__.columns
+            if isinstance(col.type, Boolean)}
+
+    @cached_classproperty
+    def all_checkgroups(cls):
+        """Returns the set of MultiChoice column names for this table."""
+        return {
+            col.name for col in cls.__table__.columns
+            if isinstance(col.type, MultiChoice)}
+
+    @cached_classproperty
+    def regform_bools(cls):
+        """Returns the set of non-admin-only Boolean columns for this table."""
+        return {
+            colname for colname in cls.all_bools
+            if colname in cls.unrestricted}
+
+    @cached_classproperty
+    def regform_checkgroups(cls):
+        """
+        Returns the set of non-admin-only MultiChoice columns for this table.
+        """
+        return {
+            colname for colname in cls.all_checkgroups
+            if colname in cls.unrestricted}
+
+    @property
+    def session(self):
+        """
+        Returns the session object which this model instance is attached to,
+        or None if this instance is not attached to a session.
+        """
+        from uber.models import Session
+        return Session.session_factory.object_session(self)
+
+    @classmethod
+    def get_field(cls, name):
+        """Returns the column object with the provided name for this model."""
+        return cls.__table__.columns[name]
+
+    def __eq__(self, m):
+        return self.id is not None and \
+            isinstance(m, MagModel) and self.id == m.id
+
+    def __ne__(self, m):
+        return not (self == m)
+
+    def __hash__(self):
+        return hash(self.id)
+
+    @property
+    def is_new(self):
+        """
+        Boolean property indicating whether or not this instance has already
+        been saved to the database or if it's a new instance which has never
+        been saved and thus has no corresponding row in its database table.
+        """
+        return not instance_state(self).persistent
+
+    @property
+    def created(self):
+        return self.get_tracking_by_instance(
+            self, action=c.CREATED, last_only=True)
+
+    @property
+    def last_updated(self):
+        return self.get_tracking_by_instance(
+            self, action=c.UPDATED, last_only=True)
+
+    @property
+    def db_id(self):
+        """
+        A common convention in our forms is to pass an "id" parameter of "None"
+        for new objects and to pass the actual id for objects which already
+        exist in our database, which lets the backend know whether to perform a
+        save or an update.  This method returns "None" for new objects and the
+        id for existing objects, for use in such forms.
+        """
+        return None if self.is_new else self.id
+
+    def orig_value_of(self, name):
+        """
+        Sometimes we mutate a model instance but then want to get the original
+        value of a column before we changed it before we perform a save.  This
+        method returns the original value (i.e. the value currently in the db)
+        for the column whose name is provided.  If the value has not changed,
+        this just returns the current value of that field.
+        """
+        hist = get_history(self, name)
+        return (hist.deleted or hist.unchanged or [getattr(self, name)])[0]
+
+    @suffix_property
+    def _ints(self, name, val):
+        """
+        Given a column that uses a tuple of integers and strings, returns a
+        list of integers. This allows us to use 'x in y' searching for
+        MultiChoice columns.
+
+        These arguments are supplied by the @suffix_property decorator based
+        on the variable name preceding '_ints'.
+
+        Args:
+            name: The name of the column we're inspecting, e.g., "interests".
+            val: The list of tuples the column uses as possible values,
+                e.g., "c.INTEREST_OPTS".
+
+        Returns:
+            A list of integers or an empty list if val is falsey.
+
+        """
+        if not val or not name:
+            return []
+
+        choices = dict(self.get_field(name).type.choices)
+        return [int(i) for i in str(val).split(',') if int(i) in choices]
+
+    @suffix_property
+    def _label(self, name, val):
+        if not val or not name:
+            return ''
+
+        try:
+            val = int(val)
+        except ValueError:
+            log.debug(
+                '{} is not an int. Did we forget to migrate data for {} '
+                'during a DB migration?', val, name)
+            return ''
+
+        label = self.get_field(name).type.choices.get(val)
+        if not label:
+            log.debug(
+                '{} does not have a label for {}, check your enum '
+                'generating code', name, val)
+        return label
+
+    @suffix_property
+    def _local(self, name, val):
+        return val.astimezone(c.EVENT_TIMEZONE)
+
+    @suffix_property
+    def _labels(self, name, val):
+        ints = getattr(self, name + '_ints')
+        labels = dict(self.get_field(name).type.choices)
+        return sorted(labels[i] for i in ints)
+
+    def __getattr__(self, name):
+        suffixed = suffix_property.check(self, name)
+        if suffixed is not None:
+            return suffixed
+
+        choice = getattr(c, name, None)
+        if choice is not None:
+            if len(self.multichoice_columns) == 1:
+                multi = self.multichoice_columns[0]
+                if choice in multi.type.choices_dict:
+                    return choice in getattr(self, multi.name + '_ints')
+
+        if name.startswith('is_'):
+            return self.__class__.__name__.lower() == name[3:]
+
+        raise AttributeError(self.__class__.__name__ + '.' + name)
+
+    def get_tracking_by_instance(self, instance, action, last_only=True):
+        from uber.models.tracking import Tracking
+        query = self.session.query(Tracking).filter_by(
+            fk_id=instance.id, action=action).order_by(Tracking.when.desc())
+        return query.first() if last_only else query.all()
+
+    def apply(
+            self, params, *, bools=(), checkgroups=(), restricted=True,
+            ignore_csrf=True):
+        """
+        Args:
+            restricted (bool): If True, restrict any changes only to fields
+                which we allow attendees to set on their own. If False, allow
+                changes to any fields.
+        """
+        bools = self.regform_bools if restricted else bools
+        checkgroups = self.regform_checkgroups if restricted else checkgroups
+        for column in self.__table__.columns:
+            if (not restricted or column.name in self.unrestricted) and \
+                    column.name in params and column.name != 'id':
+
+                if isinstance(params[column.name], list):
+                    value = ','.join(map(str, params[column.name]))
+                elif isinstance(params[column.name], bool):
+                    value = params[column.name]
+                elif params[column.name] is None:
+                    value = None
+                else:
+                    value = str(params[column.name]).strip()
+
+                try:
+                    if value is None:
+                        pass  # Totally fine for value to be None
+                    elif isinstance(column.type, Float):
+                        if value == '':
+                            value = None
+                        else:
+                            value = float(value)
+                    elif isinstance(column.type, Numeric):
+                        if value == '':
+                            value = None
+                        elif value.endswith('.0'):
+                            value = int(value[:-2])
+                    elif isinstance(column.type, (Choice, Integer)):
+                        if value == '':
+                            value = None
+                        else:
+                            value = int(float(value))
+                    elif isinstance(column.type, UTCDateTime):
+                        try:
+                            value = datetime.strptime(
+                                value, c.TIMESTAMP_FORMAT)
+                        except ValueError:
+                            value = dateparser.parse(value)
+                        value = c.EVENT_TIMEZONE.localize(value)
+                    elif isinstance(column.type, Date):
+                        try:
+                            value = datetime.strptime(value, c.DATE_FORMAT)
+                        except ValueError:
+                            value = dateparser.parse(value)
+                        value = value.date()
+                except Exception as error:
+                    log.debug(
+                        'Ignoring error coercing value for column {}.{}: {}',
+                        self.__tablename__, column.name, error)
+
+                setattr(self, column.name, value)
+
+        for column in self.__table__.columns:
+            if (not restricted or column.name in self.unrestricted) and (
+                    column.type is JSON or isinstance(column.type, JSON)):
+                fields = getattr(self, '_{}_fields'.format(column.name), {})
+                for field in fields.keys():
+                    if field in params:
+                        setattr(self, field, params[field])
+
+        if cherrypy.request.method.upper() == 'POST':
+            for column in self.__table__.columns:
+                if column.name in bools:
+                    setattr(
+                        self,
+                        column.name,
+                        bool(int(params.get(column.name, 0))))
+                elif column.name in checkgroups and column.name not in params:
+                    setattr(self, column.name, '')
+
+            if not ignore_csrf:
+                check_csrf(params.get('csrf_token'))
+
+        return self
+
+    def timespan(self, minute_increment=60):
+        def minutestr(dt):
+            return ':30' if dt.minute == 30 else ''
+
+        timespan = timedelta(minutes=minute_increment * self.duration)
+        endtime = self.start_time_local + timespan
+
+        startstr = self.start_time_local.strftime('%I').lstrip('0') + \
+            minutestr(self.start_time_local)
+
+        endstr = endtime.strftime('%I').lstrip('0') + \
+            minutestr(endtime) + endtime.strftime('%p').lower()
+
+        if self.start_time_local.day == endtime.day:
+            endstr += endtime.strftime(' %A')
+            if self.start_time_local.hour < 12 and endtime.hour >= 12:
+                return startstr + 'am - ' + endstr
+            else:
+                return startstr + '-' + endstr
+        else:
+            return startstr + self.start_time_local.strftime('pm %a - ') + \
+                endstr + endtime.strftime(' %a')
+
+
+# Make all of our model classes available from uber.models
+from uber.models.email import *  # noqa: F401,E402,F403
+from uber.models.core import *  # noqa: F401,E402,F403
+from uber.models.commerce import *  # noqa: F401,E402,F403
+from uber.models.promo_code import *  # noqa: F401,E402,F403
+from uber.models.admin import *  # noqa: F401,E402,F403
+from uber.models.tracking import *  # noqa: F401,E402,F403
+from uber.models.types import *  # noqa: F401,E402,F403
+
+# Exlplicitly import models used by the Session class to quiet flake8
+from uber.models.email import Email  # noqa: E402
+from uber.models.core import Attendee, Group  # noqa: E402
+from uber.models.admin import AdminAccount, Job, Shift, WatchList  # noqa: E402
+from uber.models.tracking import track_changes_after_flush  # noqa: E402
+
+
+class Session(SessionManager):
+    # This looks strange, but `sqlalchemy.create_engine` will throw an error
+    # if it's passed arguments that aren't supported by the given DB engine.
+    # For example, SQLite doesn't support either `pool_size` or `max_overflow`,
+    # so if `sqlalchemy_pool_size` or `sqlalchemy_max_overflow` are set with
+    # a value of -1, they are not added to the keyword args.
+    _engine_kwargs = dict((k, v) for (k, v) in [
+        ('pool_size', c.SQLALCHEMY_POOL_SIZE),
+        ('max_overflow', c.SQLALCHEMY_MAX_OVERFLOW)] if v > -1)
+    engine = sqlalchemy.create_engine(c.SQLALCHEMY_URL, **_engine_kwargs)
+
+    @classmethod
+    def initialize_db(cls, modify_tables=False, drop=False, initialize=False):
+        """
+        Initialize the database and optionally create/drop tables.
+
+        Initializes the database connection for use, and attempt to create any
+        tables registered in our metadata which do not actually exist yet in
+        the database.
+
+        This calls the underlying sideboard function, HOWEVER, in order to
+        actually create any tables, you must specify modify_tables=True.  The
+        reason is, we need to wait for all models from all plugins to insert
+        their mixin data, so we wait until one spot in order to create the
+        database tables.
+
+        Any calls to initialize_db() that do not specify modify_tables=True or
+        drop=True are ignored.
+
+        i.e. anywhere in Sideboard that calls initialize_db() will be ignored.
+        i.e. ubersystem is forcing all calls that don't specify
+        modify_tables=True or drop=True to be ignored.
+
+        Calling initialize_db with modify_tables=False and drop=True will leave
+        you with an empty database.
+
+        Keyword Arguments:
+            modify_tables: If False, this function will not attempt to create
+                any database objects (tables, columns, constraints, etc...)
+                Defaults to False.
+            drop: USE WITH CAUTION: If True, then we will drop any tables in
+                the database. Defaults to False.
+        """
+        if drop or modify_tables or initialize:
+            super(Session, cls).initialize_db(drop=drop, create=modify_tables)
+            if drop:
+                from uber.migration import stamp
+                stamp('heads' if modify_tables else None)
+
+    class QuerySubclass(Query):
+        @property
+        def is_single_table_query(self):
+            return len(self.column_descriptions) == 1
+
+        @property
+        def model(self):
+            assert self.is_single_table_query, (
+                'actions such as .order() and .icontains() and .iexact() are '
+                'only valid for single-table queries')
+
+            return self.column_descriptions[0]['type']
+
+        def order(self, attrs):
+            order = []
+            for attr in listify(attrs):
+                col = getattr(self.model, attr.lstrip('-'))
+                order.append(col.desc() if attr.startswith('-') else col)
+            return self.order_by(*order)
+
+        def icontains_condition(self, attr=None, val=None, **filters):
+            """
+            Take column names and values, and build a condition/expression
+            that is true when all named columns contain the corresponding
+            values, case-insensitive.
+
+            This operation is very similar to the "contains" method in
+            SQLAlchemy, but case insensitive - i.e. it uses "ilike" instead
+            of "like".
+
+            Note that an "and" is used: all columns must match, not just one.
+            More complex conditions can be built by using or_/etc on the result
+            of this method.
+            """
+            conditions = []
+            if len(self.column_descriptions) == 1 and filters:
+                for colname, val in filters.items():
+                    conditions.append(
+                        getattr(self.model, colname).ilike('%{}%'.format(val)))
+            if attr and val:
+                conditions.append(attr.ilike('%{}%'.format(val)))
+            return and_(*conditions)
+
+        def icontains(self, attr=None, val=None, **filters):
+            """
+            Take the names of columns and values, and filters the query to
+            items where each named columns contain the values,
+            case-insensitive.
+
+            This operation is very similar to calling
+            query.filter(contains(...)), but works with a case-insensitive
+            "contains".
+
+            Note that an "and" is used: all columns must match, not just one.
+            """
+            condition = self.icontains_condition(attr=attr, val=val, **filters)
+            return self.filter(condition)
+
+        def iexact(self, **filters):
+            filters = [
+                func.lower(getattr(self.model, attr)) == func.lower(val)
+                for attr, val in filters.items()]
+            return self.filter(*filters)
+
+    class SessionMixin:
+        def admin_attendee(self):
+            return self.admin_account(cherrypy.session['account_id']).attendee
+
+        def logged_in_volunteer(self):
+            return self.attendee(cherrypy.session['staffer_id'])
+
+        def checklist_status(self, slug, department):
+            attendee = self.admin_attendee()
+            conf = DeptChecklistConf.instances.get(slug)
+            if not conf:
+                raise ValueError(
+                    "Can't access dept checklist INI settings for section "
+                    "'{}', check your INI file".format(slug))
+
+            is_relevant = attendee.is_single_dept_head and \
+                attendee.assigned_depts_ints == [int(department or 0)]
+            return {
+                'conf': conf,
+                'relevant': is_relevant,
+                'completed': conf.completed(attendee)
+            }
+
+        def jobs_for_signups(self):
+            fields = [
+                'name', 'location_label', 'description', 'weight',
+                'start_time_local', 'end_time_local', 'duration',
+                'weighted_hours', 'restricted', 'extra15', 'taken']
+            jobs = self.logged_in_volunteer().possible_and_current
+            restricted_hours = set()
+            for job in jobs:
+                if job.restricted:
+                    restricted_hours.add(frozenset(job.hours))
+            return [
+                job.to_dict(fields)
+                for job in jobs
+                if (job.restricted or
+                    frozenset(job.hours) not in restricted_hours)]
+
+        def guess_attendee_watchentry(self, attendee):
+            or_clauses = [
+                func.lower(WatchList.first_names).contains(
+                    attendee.first_name.lower()),
+                and_(
+                    WatchList.email != '',
+                    func.lower(WatchList.email) == attendee.email.lower())]
+
+            if attendee.birthdate:
+                if isinstance(attendee.birthdate, six.string_types):
+                    try:
+                        birthdate = dateparser.parse(attendee.birthdate).date()
+                    except:
+                        log.debug(
+                            'Error parsing attendee birthdate: '
+                            '{}'.format(attendee.birthdate))
+                    else:
+                        or_clauses.append(WatchList.birthdate == birthdate)
+                elif isinstance(attendee.birthdate, datetime):
+                    or_clauses.append(
+                        WatchList.birthdate == attendee.birthdate.date())
+                elif isinstance(attendee.birthdate, date):
+                    or_clauses.append(
+                        WatchList.birthdate == attendee.birthdate)
+
+            return self.query(WatchList).filter(and_(
+                or_(*or_clauses),
+                func.lower(WatchList.last_name) == attendee.last_name.lower(),
+                WatchList.active == True)).all()  # noqa: E712
+
+        def get_account_by_email(self, email):
+            return self.query(AdminAccount) \
+                .join(Attendee) \
+                .filter(func.lower(Attendee.email) == func.lower(email)).one()
+
+        def no_email(self, subject):
+            return not self.query(Email).filter_by(subject=subject).all()
+
+        def lookup_attendee(self, first_name, last_name, email, zip_code):
+            attendee = self.query(Attendee).iexact(
+                first_name=first_name,
+                last_name=last_name,
+                email=email,
+                zip_code=zip_code).filter(
+                    Attendee.badge_status != c.INVALID_STATUS).first()
+
+            if attendee:
+                return attendee
+
+            raise ValueError('Attendee not found')
+
+        def add_promo_code_to_attendee(self, attendee, code):
+            """
+            Convenience method for adding a promo code to an attendee.
+
+            This method sets both the `promo_code` and `promo_code_id`
+            properties of `attendee`. Due to the way the `Attendee.promo_code`
+            relationship is defined, the `Attendee.promo_code_id` isn't
+            automatically set, which makes this method a nice way of setting
+            both.
+
+            Arguments:
+                attendee (Attendee): The Attendee for which the promo code
+                    should be added.
+                code (str): The promo code as typed by an end user, or an
+                    empty string to unset the promo code.
+
+            Returns:
+                str: Either a failure message or an empty string
+                    indicating success.
+            """
+            code = code.strip() if code else ''
+            if code:
+                attendee.promo_code = self.lookup_promo_code(code)
+                if attendee.promo_code:
+                    attendee.promo_code_id = attendee.promo_code.id
+                    return ''
+                else:
+                    attendee.promo_code_id = None
+                    return 'The promo code you entered is invalid.'
+            else:
+                attendee.promo_code = None
+                attendee.promo_code_id = None
+                return ''
+
+        def lookup_promo_code(self, code):
+            """
+            Convenience method for finding a promo code by id or code.
+
+            Arguments:
+                code (str): The id or code to search for.
+
+            Returns:
+                PromoCode: Either the matching PromoCode object, or None
+                    if not found.
+            """
+            from uber.models.promo_code import PromoCode
+            if isinstance(code, uuid.UUID):
+                code = code.hex
+
+            normalized_code = PromoCode.normalize_code(code)
+            if not normalized_code:
+                return None
+
+            unambiguous_code = PromoCode.disambiguate_code(code)
+            clause = or_(
+                PromoCode.normalized_code == normalized_code,
+                PromoCode.normalized_code == unambiguous_code)
+
+            # Make sure that code is a valid UUID before adding
+            # PromoCode.id to the filter clause
+            try:
+                promo_code_id = uuid.UUID(normalized_code).hex
+            except:
+                pass
+            else:
+                clause = clause.or_(PromoCode.id == promo_code_id)
+
+            return self.query(PromoCode).filter(clause).order_by(
+                PromoCode.normalized_code.desc()).first()
+
+        def get_next_badge_num(self, badge_type):
+            """
+            Returns the next badge available for a given badge type. This is
+            essentially a wrapper for auto_badge_num that accounts for new or
+            changed objects in the session.
+
+            Args:
+                badge_type: Used to pass to auto_badge_num and to ignore
+                    objects in the session that aren't within the badge
+                    type's range.
+
+            """
+            badge_type = get_real_badge_type(badge_type)
+
+            new_badge_num = self.auto_badge_num(badge_type)
+            lower_bound = c.BADGE_RANGES[badge_type][0]
+            upper_bound = c.BADGE_RANGES[badge_type][1]
+
+            # Adjusts the badge number based on badges in the session
+            all_models = chain(self.new, self.dirty)
+            for attendee in [m for m in all_models if isinstance(m, Attendee)]:
+                if attendee.badge_num is not None and \
+                        lower_bound <= attendee.badge_num <= upper_bound:
+                    new_badge_num = max(new_badge_num, 1 + attendee.badge_num)
+
+            assert new_badge_num < upper_bound, \
+                'There are no more badge numbers available in this range!'
+
+            return new_badge_num
+
+        def update_badge(self, attendee, old_badge_type, old_badge_num):
+            """
+            This should be called whenever an attendee's badge type or badge
+            number is changed. It checks if the attendee will still require a
+            badge number with their new badge type, and if so, sets their
+            number to either the number specified by the admin or the lowest
+            available badge number in that range.
+
+            Args:
+                attendee: The Attendee() object whose badge is being changed.
+                old_badge_type: The old badge type.
+                old_badge_num: The old badge number.
+
+            """
+            from uber.badge_funcs import needs_badge_num
+
+            if c.SHIFT_CUSTOM_BADGES and c.BEFORE_PRINTED_BADGE_DEADLINE and \
+                    not c.AT_THE_CON:
+
+                badge_collision = False
+                if attendee.badge_num:
+                    badge_collision = self.query(Attendee.badge_num).filter(
+                        Attendee.badge_num == attendee.badge_num,
+                        Attendee.id != attendee.id).first()
+
+                desired_badge_num = attendee.badge_num
+                if old_badge_num:
+                    if attendee.badge_num and badge_collision:
+                        if old_badge_type == attendee.badge_type:
+                            if old_badge_num < attendee.badge_num:
+                                self.shift_badges(
+                                    old_badge_type,
+                                    old_badge_num + 1,
+                                    until=attendee.badge_num,
+                                    down=True)
+                            else:
+                                self.shift_badges(
+                                    old_badge_type,
+                                    attendee.badge_num,
+                                    until=old_badge_num - 1,
+                                    up=True)
+                        else:
+                            self.shift_badges(
+                                old_badge_type,
+                                old_badge_num + 1,
+                                down=True)
+                            self.shift_badges(
+                                attendee.badge_type,
+                                attendee.badge_num,
+                                up=True)
+                    else:
+                        self.shift_badges(
+                            old_badge_type, old_badge_num + 1, down=True)
+
+                elif attendee.badge_num and badge_collision:
+                    self.shift_badges(
+                        attendee.badge_type, attendee.badge_num, up=True)
+
+                attendee.badge_num = desired_badge_num
+
+            if not attendee.badge_num and needs_badge_num(attendee):
+                attendee.badge_num = self.get_next_badge_num(
+                    attendee.badge_type)
+
+            return 'Badge updated'
+
+        def auto_badge_num(self, badge_type):
+            """
+            Gets the next available badge number for a badge type's range.
+
+            Plugins can override the logic here if need be without worrying
+            about handling dirty sessions.
+
+            Args:
+                badge_type: Used as a starting point if no badges of the same
+                    type exist, and to select badges within a specific range.
+
+            """
+            in_range = self.query(Attendee.badge_num).filter(
+                Attendee.badge_num != None,  # noqa: E711
+                Attendee.badge_num >= c.BADGE_RANGES[badge_type][0],
+                Attendee.badge_num <= c.BADGE_RANGES[badge_type][1])
+
+            in_range_list = [
+                int(row[0]) for row in in_range.order_by(Attendee.badge_num)]
+
+            if len(in_range_list):
+                # Searches badge range for a gap in badge numbers; if none
+                # found, returns the latest badge number + 1.
+                # Doing this lets admins manually set high badge numbers
+                # without filling up the badge type's range.
+                start, end = c.BADGE_RANGES[badge_type][0], in_range_list[-1]
+                gap_nums = sorted(
+                    set(range(start, end + 1)).difference(in_range_list))
+
+                if not gap_nums:
+                    return end + 1
+                else:
+                    return gap_nums[0]
+            else:
+                return c.BADGE_RANGES[badge_type][0]
+
+        def shift_badges(
+                self, badge_type, badge_num, *, until=None, up=False,
+                down=False):
+
+            if not c.SHIFT_CUSTOM_BADGES or c.AFTER_PRINTED_BADGE_DEADLINE or \
+                    c.AT_THE_CON:
+                return False
+
+            from uber.badge_funcs import get_badge_type
+            (calculated_badge_type, error) = get_badge_type(badge_num)
+            badge_type = calculated_badge_type or badge_type
+            until = until or c.BADGE_RANGES[badge_type][1]
+
+            shift = 1 if up else -1
+            query = self.query(Attendee).filter(
+                Attendee.badge_num != None,  # noqa: E711
+                Attendee.badge_num >= badge_num,
+                Attendee.badge_num <= until)
+
+            query.update(
+                {Attendee.badge_num: Attendee.badge_num + shift},
+                synchronize_session='evaluate')
+
+            return True
+
+        def valid_attendees(self):
+            return self.query(Attendee).filter(
+                Attendee.badge_status != c.INVALID_STATUS)
+
+        def attendees_with_badges(self):
+            return self.query(Attendee).filter(not_(Attendee.badge_status.in_(
+                [c.INVALID_STATUS, c.REFUNDED_STATUS, c.DEFERRED_STATUS])))
+
+        def all_attendees(self, only_staffing=False):
+            """
+            Returns a Query of Attendees with efficient loading for groups and
+            shifts/jobs.
+
+            In some cases we only want to return attendees where "staffing"
+            is true, because before the event people can't sign up for shifts
+            unless they're marked as volunteers.  However, on-site we relax
+            that restriction, so we'll get attendees with shifts who are not
+            actually marked as staffing.  We therefore have an optional
+            parameter for clients to indicate that all attendees should be
+            returned.
+            """
+            staffing_filter = [Attendee.staffing == True] \
+                if only_staffing else []  # noqa: E712
+
+            badge_filter = Attendee.badge_status.in_(
+                [c.NEW_STATUS, c.COMPLETED_STATUS])
+
+            return self.query(Attendee) \
+                .filter(badge_filter, *staffing_filter) \
+                .options(
+                    subqueryload(Attendee.group),
+                    subqueryload(Attendee.shifts).subqueryload(Shift.job)) \
+                .order_by(Attendee.full_name)
+
+        def staffers(self):
+            return self.all_attendees(only_staffing=True)
+
+        def jobs(self, location=None):
+            return self.query(Job) \
+                .filter_by(**{'location': location} if location else {}) \
+                .options(
+                    subqueryload(Job.shifts)
+                    .subqueryload(Shift.attendee)
+                    .subqueryload(Attendee.group)) \
+                .order_by(Job.start_time, Job.name)
+
+        def staffers_for_dropdown(self):
+            query = self.query(Attendee.id, Attendee.full_name)
+            return [{
+                'id': id,
+                'full_name': full_name.title()}
+                for id, full_name in query.filter_by(staffing=True)
+                                          .order_by(Attendee.full_name)]
+
+        def single_dept_heads(self, dept=None):
+            assigned = {'assigned_depts': str(dept)} if dept else {}
+            return self.query(Attendee) \
+                .filter(Attendee.ribbon.contains(c.DEPT_HEAD_RIBBON)) \
+                .filter_by(**assigned) \
+                .order_by(Attendee.full_name).all()
+
+        def match_to_group(self, attendee, group):
+            available = [a for a in group.attendees if a.is_unassigned]
+            if not available:
+                return 'The last badge for that group has already been ' \
+                    'assigned by another station'
+
+            matching = [
+                a for a in available if a.badge_type == attendee.badge_type]
+
+            if not matching:
+                return 'Badge #{} is a {} badge, but {} has no badges of ' \
+                    'that type'.format(
+                        attendee.badge_num,
+                        attendee.badge_type_label,
+                        group.name)
+            else:
+                # First preserve the attributes to copy to the new group member
+                attrs = matching[0].to_dict(attrs=[
+                    'group', 'group_id', 'paid', 'amount_paid', 'ribbon'])
+
+                # Then delete the old unassigned group member
+                self.delete(matching[0])
+
+                # Flush the deletion so the badge shifting code is performed
+                self.flush()
+
+                # Copy the attributes we preserved
+                attendee.apply(attrs, restricted=False)
+
+                # Ensure the attendee is added to the session
+                self.add(attendee)
+                self.commit()
+
+        def search(self, text, *filters):
+            attendees = self.query(Attendee).outerjoin(Attendee.group) \
+                .options(joinedload(Attendee.group)).filter(*filters)
+
+            if ':' in text:
+                target, term = text.split(':', 1)
+                if target == 'email':
+                    return attendees.icontains(
+                        Attendee.normalized_email,
+                        Attendee.normalize_email(term))
+                elif target == 'group':
+                    return attendees.icontains(Group.name, term.strip())
+
+            terms = text.split()
+            if len(terms) == 2:
+                first, last = terms
+                if first.endswith(','):
+                    last, first = first.strip(','), last
+                name_cond = attendees.icontains_condition(
+                    first_name=first, last_name=last)
+                legal_name_cond = attendees.icontains_condition(
+                    legal_name="{}%{}".format(first, last))
+                return attendees.filter(or_(name_cond, legal_name_cond))
+
+            elif len(terms) == 1 and terms[0].endswith(','):
+                last = terms[0].rstrip(',')
+                name_cond = attendees.icontains_condition(last_name=last)
+                # Known issue: search includes first name if legal name is set
+                legal_cond = attendees.icontains_condition(legal_name=last)
+                return attendees.filter(or_(name_cond, legal_cond))
+
+            elif len(terms) == 1 and terms[0].isdigit():
+                if len(terms[0]) == 10:
+                    return attendees.filter(or_(
+                        Attendee.ec_phone == terms[0],
+                        Attendee.cellphone == terms[0]))
+                elif int(terms[0]) <= sorted(
+                        c.BADGE_RANGES.items(),
+                        key=lambda badge_range: badge_range[1][0])[-1][1][1]:
+                    return attendees.filter(Attendee.badge_num == terms[0])
+
+            elif len(terms) == 1 and re.match(
+                    '^[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-'
+                    '[a-z0-9]{4}-[a-z0-9]{12}$',
+                    terms[0]):
+                return attendees.filter(or_(
+                    Attendee.id == terms[0],
+                    Attendee.public_id == terms[0],
+                    Group.id == terms[0],
+                    Group.public_id == terms[0]))
+
+            elif len(terms) == 1 and terms[0].startswith(c.EVENT_QR_ID):
+                search_uuid = terms[0][len(c.EVENT_QR_ID):]
+                if re.match(
+                        '^[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-'
+                        '[a-z0-9]{4}-[a-z0-9]{12}$',
+                        search_uuid):
+                    return attendees.filter(or_(
+                        Attendee.public_id == search_uuid,
+                        Group.public_id == search_uuid))
+
+            checks = [Group.name.ilike('%' + text + '%')]
+            check_attrs = [
+                'first_name', 'last_name', 'legal_name', 'badge_printed_name',
+                'email', 'comments', 'admin_notes', 'for_review']
+
+            for attr in check_attrs:
+                checks.append(getattr(Attendee, attr).ilike('%' + text + '%'))
+            return attendees.filter(or_(*checks))
+
+        def delete_from_group(self, attendee, group):
+            """
+            Sometimes we want to delete an attendee badge which is part of a
+            group.  In most cases, we could just say "session.delete(attendee)"
+            but sometimes we need to make sure that the attendee is ALSO
+            removed from the "group.attendees" list before we commit, since the
+            number of attendees in a group is used in our presave_adjustments()
+            code to update the group price.  So anytime we delete an attendee
+            in a group, we should use this method.
+            """
+            self.delete(attendee)
+            group.attendees.remove(attendee)
+
+        def assign_badges(
+                self, group, new_badge_count, new_badge_type=c.ATTENDEE_BADGE,
+                new_ribbon_type=None, paid=c.PAID_BY_GROUP,
+                **extra_create_args):
+
+            diff = int(new_badge_count) - group.badges
+            sorted_unassigned = sorted(
+                group.floating, key=lambda a: a.registered, reverse=True)
+
+            ribbon_to_use = ','.join(map(str, listify(new_ribbon_type))) \
+                if new_ribbon_type else group.new_ribbon
+
+            if int(new_badge_type) in c.PREASSIGNED_BADGE_TYPES and \
+                    c.AFTER_PRINTED_BADGE_DEADLINE and diff > 0:
+                return 'Custom badges have already been ordered, so you ' \
+                    'will need to select a different badge type'
+            elif diff > 0:
+                for i in range(diff):
+                    group.attendees.append(Attendee(
+                        badge_type=new_badge_type,
+                        ribbon=ribbon_to_use,
+                        paid=paid,
+                        **extra_create_args))
+            elif diff < 0:
+                if len(group.floating) < abs(diff):
+                    return 'You cannot reduce the number of badges for a ' \
+                        'group to below the number of assigned badges'
+                else:
+                    for attendee in sorted_unassigned[:abs(diff)]:
+                        self.delete_from_group(attendee, group)
+
+        def assign(self, attendee_id, job_id):
+            """
+            assign an Attendee to a Job by creating a Shift
+            :return: 'None' on success, error message on failure
+            """
+            job = self.job(job_id)
+            attendee = self.attendee(attendee_id)
+
+            if job.restricted and not attendee.trusted_in(job.location):
+                return 'You cannot assign an attendee who is not trusted ' \
+                    'in this department to a restricted shift'
+
+            if job.slots <= len(job.shifts):
+                return 'All slots for this job have already been filled'
+
+            if not job.no_overlap(attendee):
+                return 'This volunteer is already signed up for a shift ' \
+                    'during that time'
+
+            self.add(Shift(attendee=attendee, job=job))
+            self.commit()
+
+        def affiliates(self):
+            amounts = defaultdict(
+                int, {a: -i for i, a in enumerate(c.DEFAULT_AFFILIATES)})
+
+            query = self.query(Attendee.affiliate, Attendee.amount_extra) \
+                .filter(and_(
+                    Attendee.amount_extra >
+                    0, Attendee.affiliate != ''))
+
+            for aff, amt in query:
+                amounts[aff] += amt
+
+            return [{
+                'id': aff,
+                'text': aff,
+                'total': max(0, amt)
+            } for aff, amt in sorted(amounts.items(), key=lambda tup: -tup[1])]
+
+        def insert_test_admin_account(self):
+            """
+            Insert a test admin into the database with username
+            "magfest@example.com" password "magfest" this is ONLY allowed if
+            no other admins already exist in the database.
+
+            Returns:
+                bool: True if success, False if failure
+            """
+            if self.query(AdminAccount).count() != 0:
+                return False
+
+            attendee = Attendee(
+                placeholder=True,
+                first_name='Test',
+                last_name='Developer',
+                email='magfest@example.com',
+                badge_type=c.ATTENDEE_BADGE,
+            )
+            self.add(attendee)
+
+            self.add(AdminAccount(
+                attendee=attendee,
+                access=','.join(str(level) for level, name in c.ACCESS_OPTS),
+                hashed=bcrypt.hashpw('magfest', bcrypt.gensalt())
+            ))
+
+            return True
+
+        def bulk_insert(self, models):
+            """
+            Convenience method for bulk inserting model objects.
+
+            In general, doing a bulk insert is much faster than individual
+            inserts, but the whole insert will fail if a single object
+            violates the database's referential integrity.
+
+            This function does a bulk insert, but if an `IntegrityError` is
+            encountered, it falls back to inserting the model objects
+            one-by-one, and ignores the individual integrity errors.
+
+            Arguments:
+                models (list): A list of sqlalchemy model objects.
+
+            Returns:
+                list: A list of model objects that was succesfully inserted.
+                    The returned list will not include any model objects that
+                    failed insertion.
+            """
+            for model in models:
+                model.presave_adjustments()
+            try:
+                self.bulk_save_objects(models)
+                self.commit()
+                return models
+            except IntegrityError as error:
+                log.debug('Bulk insert failed: {}', error)
+                self.rollback()
+
+                # Bulk insert failed, so insert one at a time and ignore errors
+                inserted_models = []
+                for model in models:
+                    try:
+                        self.add(model)
+                        self.commit()
+                        inserted_models.append(model)
+                    except IntegrityError:
+                        log.debug('Individual insert failed: {}', error)
+                        # Ignore db integrity errors
+                        self.rollback()
+                return inserted_models
+
+    @classmethod
+    def model_mixin(cls, model):
+        if model.__name__ in ['SessionMixin', 'QuerySubclass']:
+            target = getattr(cls, model.__name__)
+        else:
+            for target in cls.all_models():
+                if target.__name__ == model.__name__:
+                    break
+            else:
+                raise ValueError(
+                    'No existing model with name {}'.format(model.__name__))
+
+        for name in dir(model):
+            if not name.startswith('_'):
+                attr = getattr(model, name)
+                if hasattr('target', '__table__') and \
+                        name in target.__table__.c:
+                    attr.key = attr.key or name
+                    attr.name = attr.name or name
+                    attr.table = target.__table__
+                    target.__table__.c.replace(attr)
+                else:
+                    setattr(target, name, attr)
+        return target
+
+
+def _make_getter(model):
+    def getter(
+            self, params=None, *, bools=(), checkgroups=(), allowed=(),
+            restricted=False, ignore_csrf=False, **query):
+
+        if query:
+            return self.query(model).filter_by(**query).one()
+        elif isinstance(params, str):
+            return self.query(model).filter_by(id=params).one()
+        else:
+            params = params.copy()
+            id = params.pop('id', 'None')
+            if id == 'None':
+                inst = model()
+            else:
+                inst = self.query(model).filter_by(id=id).one()
+
+            if not ignore_csrf:
+                assert (
+                    not {k for k in params if k not in allowed} or
+                    cherrypy.request.method == 'POST'), 'POST required'
+
+            inst.apply(
+                params, bools=bools, checkgroups=checkgroups,
+                restricted=restricted, ignore_csrf=ignore_csrf)
+
+            return inst
+    return getter
+
+
+@on_startup(priority=1)
+def initialize_db(modify_tables=False):
+    """
+    Initialize the session on startup.
+
+    We want to do this only after all other plugins have had a chance to
+    initialize and add their 'mixin' data (i.e. extra colums) into the models.
+
+    Also, it's possible that the DB is still initializing and isn't ready to
+    accept connections, so, if this fails, keep trying until we're able to
+    connect.
+
+    This should be the ONLY spot (except for maintenance tools) in all of core
+    ubersystem or any plugins that attempts to create tables by passing
+    drop=True or modify_tables=True or initialize=True to
+    Session.initialize_db()
+    """
+    for _model in Session.all_models():
+        setattr(
+            Session.SessionMixin, _model.__tablename__, _make_getter(_model))
+
+    num_tries_remaining = 10
+    while not stopped.is_set():
+        try:
+            Session.initialize_db(modify_tables=modify_tables, initialize=True)
+        except KeyboardInterrupt:
+            log.critical(
+                'DB initialize: Someone hit Ctrl+C while we were starting up')
+        except:
+            num_tries_remaining -= 1
+            if num_tries_remaining == 0:
+                log.error(
+                    "DB initialize: couldn't connect to DB, we're giving up")
+                raise
+            log.error(
+                "DB initialize: can't connect to / initialize DB, will try "
+                "again in 5 seconds", exc_info=True)
+            stopped.wait(5)
+        else:
+            break
+
+
+def _presave_adjustments(session, context, instances='deprecated'):
+    for model in chain(session.dirty, session.new):
+        model.presave_adjustments()
+    for model in session.deleted:
+        model.predelete_adjustments()
+
+
+def register_session_listeners():
+    """
+    The order in which we register these listeners matters.
+    """
+    listen(Session.session_factory, 'before_flush', _presave_adjustments)
+    listen(Session.session_factory, 'after_flush', track_changes_after_flush)
+
+
+register_session_listeners()
